ml_model/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
PyMuPDF
python-docx
python-multipart
pillow
pyzbar
tldextract
validators

opencv-python
pillow
pyzbar
tldextract
joblib
numpy